Wow, a whole lot
Added newest version of the opencv tracking, a method that imports a file and returns an array of point values, can take in csv in time(ms) xCoord yCoord, or .txt in the format included on the drive, the output of their javaGrinders program. Included the test videos and updated the mask image for testing the tracking program. That's all I can think of right now